Drop unused HashArray::removeDuplicates
