+++ tests/unit/HashArray/HashArrayTest.php
@@ -45,61 +45,6 @@ abstract class HashArrayTest extends \PHPUnit_Framework_TestCase {
 	 * @dataProvider instanceProvider
 	 * @param HashArray $array
 	 */
-	public function testHasElement( HashArray $array ) {
-		$array->removeDuplicates();
-
-		/**
-		 * @var Hashable $hashable
-		 */
-		foreach ( iterator_to_array( $array ) as $hashable ) {
-			$this->assertTrue( $array->hasElement( $hashable ) );
-			$this->assertTrue( $array->hasElementHash( $hashable->getHash() ) );
-			$array->removeElement( $hashable );
-			$this->assertFalse( $array->hasElement( $hashable ) );
-			$this->assertFalse( $array->hasElementHash( $hashable->getHash() ) );
-		}
-
-		$this->assertTrue( true );
-	}
-
-	/**
-	 * @dataProvider instanceProvider
-	 * @param HashArray $array
-	 */
-	public function testRemoveElement( HashArray $array ) {
-		$array->removeDuplicates();
-
-		$elementCount = $array->count();
-
-		/**
-		 * @var Hashable $element
-		 */
-		foreach ( iterator_to_array( $array ) as $element ) {
-			$this->assertTrue( $array->hasElement( $element ) );
-
-			if ( $elementCount % 2 === 0 ) {
-				$array->removeElement( $element );
-			}
-			else {
-				$array->removeByElementHash( $element->getHash() );
-			}
-
-			$this->assertFalse( $array->hasElement( $element ) );
-			$this->assertEquals( --$elementCount, $array->count() );
-		}
-
-		$element = new PropertyNoValueSnak( 42 );
-
-		$array->removeElement( $element );
-		$array->removeByElementHash( $element->getHash() );
-
-		$this->assertTrue( true );
-	}
-
-	/**
-	 * @dataProvider instanceProvider
-	 * @param HashArray $array
-	 */
 	public function testEquals( HashArray $array ) {
 		$this->assertTrue( $array->equals( $array ) );
 		$this->assertFalse( $array->equals( 42 ) );
