+++ tests/unit/HashArray/HashArrayWithoutDuplicatesTest.php
@@ -2,8 +2,10 @@
 
 namespace Wikibase\DataModel\Tests\HashArray;
 
+use Hashable;
 use Wikibase\DataModel\Fixtures\HashArrayElement;
 use Wikibase\DataModel\HashArray;
+use Wikibase\DataModel\Snak\PropertyNoValueSnak;
 
 /**
  * @covers Wikibase\DataModel\HashArray
@@ -80,15 +82,51 @@ class HashArrayWithoutDuplicatesTest extends HashArrayTest {
 	 * @dataProvider instanceProvider
 	 * @param HashArray $array
 	 */
-	public function testRemoveDuplicates( HashArray $array ) {
-		$count = count( $array );
-		$array->removeDuplicates();
+	public function testHasElement( HashArray $array ) {
+		/**
+		 * @var Hashable $hashable
+		 */
+		foreach ( iterator_to_array( $array ) as $hashable ) {
+			$this->assertTrue( $array->hasElement( $hashable ) );
+			$this->assertTrue( $array->hasElementHash( $hashable->getHash() ) );
+			$array->removeElement( $hashable );
+			$this->assertFalse( $array->hasElement( $hashable ) );
+			$this->assertFalse( $array->hasElementHash( $hashable->getHash() ) );
+		}
 
-		$this->assertCount(
-			$count,
-			$array,
-			'Count should be the same after removeDuplicates since there can be none'
-		);
+		$this->assertTrue( true );
+	}
+
+	/**
+	 * @dataProvider instanceProvider
+	 * @param HashArray $array
+	 */
+	public function testRemoveElement( HashArray $array ) {
+		$elementCount = $array->count();
+
+		/**
+		 * @var Hashable $element
+		 */
+		foreach ( iterator_to_array( $array ) as $element ) {
+			$this->assertTrue( $array->hasElement( $element ) );
+
+			if ( $elementCount % 2 === 0 ) {
+				$array->removeElement( $element );
+			}
+			else {
+				$array->removeByElementHash( $element->getHash() );
+			}
+
+			$this->assertFalse( $array->hasElement( $element ) );
+			$this->assertEquals( --$elementCount, $array->count() );
+		}
+
+		$element = new PropertyNoValueSnak( 42 );
+
+		$array->removeElement( $element );
+		$array->removeByElementHash( $element->getHash() );
+
+		$this->assertTrue( true );
 	}
 
 	/**
